+++ HISTORY.md
@@ -1,8 +1,9 @@
 ## Changelog ##
 
-### 2014-01-01 - v4.4.2 - (Dev time: 1 hours) ###
+### 2014-01-01 - v4.4.2 - (Dev time: 1.5 hours) ###
 - Fix: Deleting first x posts deletes all posts while deleting by category (#44)
 - Fix: Posts are moved to trash even if "Delete permanently" option is selected (#45)
+- Tweak: Move request processing code for deleting by custom field to addon
 
 ### 2013-12-18 - v4.4.1 - (Dev time: 0.5 hours) ###
 - Fix: Bulk Delete menu overrides other menus at the same position